reorganise settings components
