+++ src/js/components/Settings/UserSettings/FeatureSettings.vue
