+++ src/js/components/Settings/UserSettings/PerformanceSettings.vue
