+++ src/js/components/Settings/UserSettings/StyleSettings.vue
